+++ actix-service/CHANGES.md
@@ -3,10 +3,14 @@
 ## Unreleased - 2020-xx-xx
 * `Service`, other traits, and many type signatures now take the the request type as a type
   parameter instead of an associated type. [#232]
-* Upgrade `pin-project` to `1.0`.
-
+* Add `always_ready!` and `forward_ready!` macros. [#233]
+* Crate is now `no_std`. [#233]
+* Migrate pin projections to `pin-project-lite`. [#233]
+* Remove `AndThenApplyFn` and Pipeline `and_then_apply_fn`. Use the
+  `.and_then(apply_fn(...))` construction. [#233]
 
 [#232]: https://github.com/actix/actix-net/pull/232
+[#233]: https://github.com/actix/actix-net/pull/233
 
 
 ## 1.0.6 - 2020-08-09